% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_at.R
\name{angle_at}
\alias{angle_at}
\title{Angle at Vertex}
\usage{
angle_at(a = NULL, b = NULL, c = NULL)
}
\arguments{
\item{a}{vertex a}

\item{b}{vertex b}

\item{c}{vertex c}
}
\value{
angle at vertex b given triangle (a, b, c) with 3D coords. For internal use only.
}
\description{
Compute angle at vertex b given triangle (a, b, c) with 3D coords
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
\keyword{angle_at}
