% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_sp_manifold.R
\name{bayes_sp_manifold}
\alias{bayes_sp_manifold}
\title{Spatial Bayesian Hierarchical model for compact Riemannian manifolds}
\usage{
bayes_sp_manifold(
  mesh = NULL,
  lambda = NULL,
  phi = NULL,
  y = NULL,
  samp = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  a.v = NULL,
  b.v = NULL,
  lower.alpha = NULL,
  upper.alpha = NULL,
  nu = NULL,
  verbose = TRUE,
  digits = 3
)
}
\arguments{
\item{mesh}{a mesh in polygon file format imported using `vcgPlyRead()` of class `mesh3d`.}

\item{lambda}{truncated vector of eigen-values}

\item{phi}{truncated \eqn{N\times T} matrix of eigen functions}

\item{y}{a \eqn{N\times 1} vector of the observed realizations}

\item{samp}{list containing barycentric coordinates of \eqn{N} sampled irregular locations, IDs for triangle vertices}

\item{niter}{optional; number of MCMC iterations, defaults to 5000}

\item{nburn}{optional; number of MCMC burn-in, defaults to 2500}

\item{report}{optional; MCMC batch length, defaults to 100}

\item{a.v}{optional; hyperparameter for the shape of the inverse-gamma prior on total variance, defaults to 2}

\item{b.v}{optional; hyperparameter for the scale of the inverse-gamma prior on total variance, defaults to 2}

\item{lower.alpha}{optional; hyperparameter for the lower bound for the uniform prior on the length scale parameter, defaults to 0.01}

\item{upper.alpha}{optional; hyperparameter for the upper bound for the uniform prior on the length scale parameter, defaults to 30}

\item{nu}{optional; value of the smoothness/fractal parameter, defaults to 2}

\item{verbose}{logical; if TRUE prints inference}

\item{digits}{optional; precision of print output, defaults to 3 digits}
}
\value{
A list of the observed process, sampled locations, fractal parameter, posterior post burn-in samples for \eqn{\sigma^2, \tau^2, \alpha}.
}
\description{
Fits the Bayesian hierarchical model, \eqn{Y(x)=\beta_0 + Z(x) + \epsilon(x)}, \eqn{x=(x^1,x^2,x^3)\in\mathcal{M}}, where \eqn{\mathcal{M}} is a compact Riemanian manifold. Priors are \eqn{Z(x)\sim GP(0, K(\cdot;\sigma^2,\alpha))}, and \eqn{\epsilon(x)\sim N(0,\tau^2)}. It uses a collapsed Metropolis-Hastings sampler yielding posterior Markov Chain Monte Carlo (MCMC) samples.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
