% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cot.R
\name{cot}
\alias{cot}
\title{Cotangent Function}
\usage{
cot(theta = NULL)
}
\arguments{
\item{theta}{angle}
}
\value{
cotangent of the angle. For internal use only.
}
\description{
Function to calculate cotangent of an angle
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
\keyword{cot}
