% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross}
\alias{cross}
\title{Cross Product of Vectors}
\usage{
cross(a = NULL, b = NULL)
}
\arguments{
\item{a}{\eqn{3 \times 1} vector.}

\item{b}{\eqn{3 \times 1} vector.}
}
\value{
a \eqn{3 \times 1} vector. For internal use only.
}
\description{
Function for computing cross-product of two vectors.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
\keyword{cross}
