% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1legendreM.R
\name{d1legendreM}
\alias{d1legendreM}
\title{Gradients of Legendre-Matern Covariance}
\usage{
d1legendreM(nu = 1.5, alpha = 5, Pterm = NULL)
}
\arguments{
\item{nu}{optional; smoothness or fractal parameter, defaults to 1.5}

\item{alpha}{length scale parameter}

\item{Pterm}{the precomputed list of polynomial terms from `dPl()`}
}
\value{
a cross-covariance matrix of dimension, \eqn{N\times N_G}. Internal use.
}
\description{
Gradients of Legendre-Matern Covariance used for the cross covariance for posterior inference on gradients.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
