% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2legendreM0.R
\name{d2legendreM0}
\alias{d2legendreM0}
\title{Second derivative of Legendre-Matern Covariance}
\usage{
d2legendreM0(Lmax = 20, nu = 1.5, alpha = 5)
}
\arguments{
\item{Lmax}{the truncation used for the Legendre-Matern covariance}

\item{nu}{optional; smoothness or fractal parameter, defaults to 1.5}

\item{alpha}{length scale parameter}
}
\value{
a variance matrix of dimension, \eqn{1\times 1}. Internal use.
}
\description{
Second derivative of Legendre-Matern Covariance used for the variance for posterior inference on gradients.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
