% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPl.R
\name{dPl}
\alias{dPl}
\title{Derivative of Legendre Polynomials}
\usage{
dPl(d = NULL, Lmax = 30)
}
\arguments{
\item{d}{a matrix of \eqn{N\times N_G} of distances between observed locations and grid locations}

\item{Lmax}{optional; truncation used for Legendre-Matern for original process, defaults to 30}
}
\value{
a list of length `Lmax` containing the polynomial expression term
}
\description{
Function that pre-computes to derivatives of Legendre-Polynomials
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
