% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farthest_point_subsample.R
\name{farthest_point_subsample}
\alias{farthest_point_subsample}
\title{Farthest Point Sampling for Point Clouds}
\usage{
farthest_point_subsample(coords, nkeep)
}
\arguments{
\item{coords}{a list containing a dense sample of points, IDs for triangle vertices and barycentric coordinates using `sample_points_mesh()`}

\item{nkeep}{number of points required in the grid}
}
\value{
a list of sample of points, IDs for triangle vertices and barycentric coordinates of length nkeep
}
\description{
Generates a point cloud resembling a grid for the mesh.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
