% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_manifold.R
\name{gradient_manifold}
\alias{gradient_manifold}
\title{Posterior inference on gradients}
\usage{
gradient_manifold(
  mesh = NULL,
  coords = NULL,
  grid = NULL,
  model = NULL,
  lambda = NULL,
  phi = NULL,
  d = 2,
  Vfield = NULL
)
}
\arguments{
\item{mesh}{a mesh in polygon file format imported using `vcgPlyRead()` of class `mesh3d`.}

\item{coords}{list containing barycentric coordinates of \eqn{N} sampled irregular locations, IDs for triangle vertices}

\item{grid}{grid locations generated using `farthest_point_subsample()`}

\item{model}{a list containing posterior post burn-in samples of \eqn{\sigma^2} and \eqn{\alpha}}

\item{lambda}{truncated vector of eigen-values}

\item{phi}{truncated \eqn{N\times T} matrix of eigen functions}

\item{d}{optional; manifold dimension, defaults to 2}

\item{Vfield}{optional; vector field on the mesh}
}
\value{
A list of the posterior samples for the gradients and a list containing: grid points, IDs of triangle vertices and barycentric coordinates and poosterior estimate of gradients.
}
\description{
Posterior Markov Chain Monte Carlo (MCMC) samples for gradients along a pre-supplied vector field 1-1 posterior samples of \eqn{\sigma^2} and \eqn{\alpha}. If vector field is not supplied it uses the local \eqn{x}-axis for vertices to generate a smooth vector field.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
