% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_cotangent.R
\name{laplacian_cotangent}
\alias{laplacian_cotangent}
\title{Cotangent Laplacian}
\usage{
laplacian_cotangent(mesh = NULL)
}
\arguments{
\item{mesh}{a mesh in polygon file format imported using `vcgPlyRead()` of class `mesh3d`.}
}
\value{
a sparse cotangent Laplacian matrix.
}
\description{
Function to compute a sparse cotangent Laplacian that approximates the Laplace-Beltrami operator for a compact Riemannian manifold.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
\keyword{laplacian_cotangent}
