% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendre_Pl.R
\name{legendre_Pl_list}
\alias{legendre_Pl_list}
\title{Legendre Polynomials}
\usage{
legendre_Pl_list(coords = NULL, Lmax = 20)
}
\arguments{
\item{coords}{coordinates on the 2-sphere}

\item{Lmax}{truncation}
}
\value{
a list of \eqn{N\times N} matrices of length `Lmax`. Internal use.
}
\description{
The Legendre polynomial list truncated at `Lmax`.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
