% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendreM.R
\name{legendreM}
\alias{legendreM}
\title{Legendre-Matern Covariance}
\usage{
legendreM(nu = 1.5, alpha = 5, P_l = NULL)
}
\arguments{
\item{nu}{fractal parameter}

\item{alpha}{length scale parameter}

\item{P_l}{pre-computed Legendre polynomials using `legendre_Pl_list()`. List of \eqn{N\times N} matrices of length `Lmax`.}
}
\value{
a covariance matrix of dimension, \eqn{N\times N}. Internal use.
}
\description{
The truncated Legendre-Matern covariance for the 2-sphere. The truncation is automatically obtained from the polynomial list.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
