% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern_cov_mp.R
\name{matern_cov_mp}
\alias{matern_cov_mp}
\title{Matern Covariance for compact Riemanian manifolds}
\usage{
matern_cov_mp(
  mesh = NULL,
  samp = NULL,
  lambda = NULL,
  phi = NULL,
  nu = NULL,
  alpha = NULL,
  d = 2,
  D.out = FALSE,
  cor = TRUE
)
}
\arguments{
\item{mesh}{a mesh in polygon file format imported using `vcgPlyRead()` of class `mesh3d`.}

\item{samp}{list containing barycentric coordinates of \eqn{N} sampled irregular locations, IDs for triangle vertices}

\item{lambda}{truncated vector of eigen-values}

\item{phi}{truncated \eqn{N\times T} matrix of eigen functions}

\item{nu}{fractal parameter controlling process smoothness}

\item{alpha}{length scale parameter}

\item{d}{optional; manifold dimension, defaults to 2}

\item{D.out}{logical, returns the mean of the scaling used, \eqn{C_{\nu,\alpha}(x)}}

\item{cor}{logical, if TRUE returns a correlation matrix}
}
\value{
A \eqn{N\times N} covariance or correlation matrix for the realization
}
\description{
Computes the Matern covariance matrix.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
