% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_points_mesh.R
\name{sample_points_mesh}
\alias{sample_points_mesh}
\title{Scattered sample locations on a mesh}
\usage{
sample_points_mesh(mesh = NULL, npoints = NULL)
}
\arguments{
\item{mesh}{a mesh in polygon file format imported using `vcgPlyRead()` of class `mesh3d`.}

\item{npoints}{number of samples points.}
}
\value{
A list containing a matrix of npoints x 3 points, IDs of vertices forming triangles the barycentric coordinates
}
\description{
Obtains a sample of scattered locations on a triangulear mesh.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
