% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph_gradients.R
\name{sph_gradients}
\alias{sph_gradients}
\title{Posterior inference on gradients}
\usage{
sph_gradients(model = NULL, Lmax = 30, grid_xyz = NULL)
}
\arguments{
\item{model}{a list containing posterior post burn-in samples of \eqn{\sigma^2} and \eqn{\alpha}}

\item{Lmax}{optional; defaults to 30}

\item{grid_xyz}{grid locations matrix of dimension \eqn{N_G\times 3} on the 2-sphere}
}
\value{
A list of the posterior samples for gradients at the grid locations.
}
\description{
Posterior Markov Chain Monte Carlo (MCMC) samples for gradients along a pre-supplied vector field 1-1 posterior samples of \eqn{\sigma^2} and \eqn{\alpha}.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
