% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_bayes_sp.R
\name{sphere_bayes_sp}
\alias{sphere_bayes_sp}
\title{Hierarchial spatial model for the 2-sphere}
\usage{
sphere_bayes_sp(
  y = NULL,
  coords = NULL,
  niter = NULL,
  nburn = NULL,
  report = NULL,
  a.v = NULL,
  b.v = NULL,
  lower.alpha = NULL,
  upper.alpha = NULL,
  nu = 1.5,
  Lmax = 30,
  verbose = TRUE,
  digits = 3
)
}
\arguments{
\item{y}{observed process}

\item{coords}{observed locations}

\item{niter}{optional; number of MCMC iterations, defaults to 5000}

\item{nburn}{optional; number of MCMC burn-in, defaults to 2500}

\item{report}{optional; MCMC batch length, defaults to 100}

\item{a.v}{optional; hyperparameter for the shape of the inverse-gamma prior on total variance, defaults to 2}

\item{b.v}{optional; hyperparameter for the scale of the inverse-gamma prior on total variance, defaults to 2}

\item{lower.alpha}{optional; hyperparameter for the lower bound for the uniform prior on the length scale parameter, defaults to 0.01}

\item{upper.alpha}{optional; hyperparameter for the upper bound for the uniform prior on the length scale parameter, defaults to 30}

\item{nu}{optional; value of the smoothness/fractal parameter, defaults to 1.5}

\item{Lmax}{optinal; truncation for the Legendre-Matern covariance, defaults to 30}

\item{verbose}{logical; if TRUE prints inference}

\item{digits}{optional; precision of print output, defaults to 3 digits}
}
\value{
A list of the observed process, sampled locations, fractal parameter, posterior post burn-in samples for \eqn{\sigma^2, \tau^2, \alpha}.
}
\description{
Function that fits the hierarchical maodel, \eqn{Y(x)=\beta_0+Z(x)+\epsilon(x)}, \eqn{Z(x)\sim GP(0, K(\cdot;\sigma^2,\alpha))}, and \eqn{\epsilon(x)\sim N(0,\tau^2)}.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
