% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsample.R
\name{spsample}
\alias{spsample}
\title{Posterior samples of spatial effects and intercept}
\usage{
spsample(
  mesh = NULL,
  lambda = NULL,
  phi = NULL,
  y = NULL,
  samp = NULL,
  alpha = NULL,
  sig2 = NULL,
  tau2 = NULL,
  nu = NULL,
  silent = TRUE
)
}
\arguments{
\item{mesh}{a mesh in polygon file format imported using `vcgPlyRead()` of class `mesh3d`.}

\item{lambda}{truncated vector of eigen-values}

\item{phi}{truncated \eqn{N\times T} matrix of eigen functions}

\item{y}{a \eqn{N\times 1} vector of the observed realizations}

\item{samp}{list containing barycentric coordinates of \eqn{N} sampled irregular locations, IDs for triangle vertices}

\item{alpha}{posterior MCMC samples for \eqn{\alpha} from fitting `bayes_sp_manifold()`}

\item{sig2}{posterior MCMC samples for \eqn{\sigma^2} from fitting `bayes_sp_manifold()`}

\item{tau2}{posterior MCMC samples for \eqn{\tau^2} from fitting `bayes_sp_manifold()`}

\item{nu}{value of the smoothness/fractal parameter, should be same as that used for `bayes_sp_manifold()`}

\item{silent}{logical; if TRUE does not print inference}
}
\value{
A list of posterior samples for \eqn{Z(x)} and \eqn{\beta_0}.
}
\description{
Uses posterior Markov Chain Monte Carlo (MCMC) samples of \eqn{\sigma^2}, \eqn{\tau^2} and \eqn{\alpha} to obtain samples for \eqn{Z(x)} and \eqn{\beta_0}.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
