% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsample_sphere.R
\name{spsample_sphere}
\alias{spsample_sphere}
\title{Posterior samples of spatial effects and intercept}
\usage{
spsample_sphere(
  y = NULL,
  coords = NULL,
  alpha = NULL,
  sig2 = NULL,
  tau2 = NULL,
  nu = NULL,
  Lmax = 30,
  silent = TRUE
)
}
\arguments{
\item{y}{a \eqn{N\times 1} vector of the observed realizations}

\item{coords}{coordinates of \eqn{N} sampled locations}

\item{alpha}{posterior MCMC samples for \eqn{\alpha} from fitting `bayes_sphere_sp()`}

\item{sig2}{posterior MCMC samples for \eqn{\sigma^2} from fitting `bayes_sphere_sp()`}

\item{tau2}{posterior MCMC samples for \eqn{\tau^2} from fitting `bayes_sphere_sp()`}

\item{nu}{value of the smoothness/fractal parameter, should be same as that used for `bayes_sphere_sp()`}

\item{Lmax}{optional; truncation level for the Legendre-Matern covariance, defaults to 30}

\item{silent}{logical; if TRUE does not print inference}
}
\value{
A list of posterior samples for \eqn{Z(x)} and \eqn{\beta_0}.
}
\description{
Uses posterior Markov Chain Monte Carlo (MCMC) samples of \eqn{\sigma^2}, \eqn{\tau^2} and \eqn{\alpha} to obtain samples for \eqn{Z(x)} and \eqn{\beta_0} for the 2-sphere partcularly.
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
