% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_areas.R
\name{triangle_areas}
\alias{triangle_areas}
\title{Area of a triangle}
\usage{
triangle_areas(V, F)
}
\arguments{
\item{V}{\eqn{K \times 3} matrix containing homologous vertex coordinates}

\item{F}{\eqn{N_T \times 3} matrix containing vertex indices forming triangular faces}
}
\value{
a scalar for the area of a triangle. For internal use.
}
\description{
Function to compute the area of a triangle in the mesh
}
\author{
Aritra Halder <aritra.halder@drexel.edu>
}
\keyword{triangle_areas}
